+++ src/myoddweb.commandlineparser/Interfaces/IRulesOutputFormatter.cs
@@ -5,7 +5,6 @@
     /// <summary>
     /// Output the rules.
     /// </summary>
-    /// <param name="rules"></param>
-    void Write(IReadOnlyCommandlineArgumentRules rules);
+    void Write();
   }
 }