+++ src/myoddweb.commandlineparser/OutputFormatter/ConsoleOutputFormatter.cs
@@ -1,51 +1,110 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
+using System.Diagnostics;
 using System.Linq;
 using myoddweb.commandlineparser.Interfaces;
+using myoddweb.commandlineparser.Rules;
 
 namespace myoddweb.commandlineparser.OutputFormatter
 {
   public class ConsoleOutputFormatter : IRulesOutputFormatter
-  {
-    /// <inheritdoc />
-    public void Write(IReadOnlyCommandlineArgumentRules rules)
-    {
-      // the app name.
-      var fileName = Path.GetFileNameWithoutExtension(GetType().Assembly.Location);
-      var usage = $"Usage: {fileName}";
-      Console.WriteLine( usage );
-      var usageLenWithPadding = usage.Length + 2;
-      foreach (var rule in rules)
-      {
-        // build the string
-        Console.WriteLine(rule.IsRequired
-          ? $"{FlattenKeys(rule.Keys).PadLeft(usageLenWithPadding)}"
-          : $"[{FlattenKeys(rule.Keys).PadLeft(usageLenWithPadding)}]");
-      }
+  { 
+    /// <summary>
+    /// The rules we want to dispay
+    /// </summary>
+    private readonly IReadOnlyCommandlineArgumentRules _rules;
 
-      Console.WriteLine();
+    /// <summary>
+    /// The calculated max key length
+    /// </summary>
+    private int? _maxKeyLength;
 
-      // find the max key len
-      var len = FindMaxKeyLength(rules);
-      var lenWithPadding = len + 2;
-
-      // then we can output everything
-      foreach (var rule in rules)
-      {
-        // build the string
-        Console.WriteLine($"{FlattenKeys(rule.Keys).PadRight(lenWithPadding)}:{rule.Description}");
-      }
-    }
+    /// <summary>
+    /// The command line leading pattern
+    /// </summary>
+    private readonly string _leadingPattern;
 
     /// <summary>
     /// Find the longest flattened key.
     /// </summary>
-    /// <param name="rules"></param>
     /// <returns></returns>
-    private static int FindMaxKeyLength(IReadOnlyCommandlineArgumentRules rules)
+    private int MaxKeyLength
     {
-      return rules.Select(rule => FlattenKeys(rule.Keys)).Select(fullKey => fullKey.Length).Concat(new[] {0}).Max();
+      get
+      {
+        if (_maxKeyLength.HasValue)
+        {
+          return _maxKeyLength.Value;
+        }
+
+        _maxKeyLength = _rules.Select(rule => FlattenKeys(rule.Keys, false)).Select(fullKey => fullKey.Length)
+          .Concat(new[] { 0 }).Max();
+        return _maxKeyLength.Value;
+      }
+    }
+
+    public ConsoleOutputFormatter(IReadOnlyCommandlineArgumentRules rules, string leadingPattern)
+    {
+      _rules = rules ?? throw new ArgumentNullException( nameof(rules));
+      _leadingPattern = leadingPattern ?? throw new ArgumentNullException(nameof(leadingPattern));
+    }
+
+    /// <inheritdoc />
+    public void Write()
+    {
+      WriteUsage();
+      Console.WriteLine();
+
+      // find the max key len
+      var lenWithPadding = MaxKeyLength + 2;
+
+      // then we can output everything
+      foreach (var rule in _rules)
+      {
+        // build the string
+        Console.WriteLine($"{FlattenKeys(rule.Keys, false ).PadRight(lenWithPadding)}:{rule.Description}");
+      }
+    }
+
+    /// <summary>
+    /// Output the help usage.
+    /// </summary>
+    private void WriteUsage()
+    {
+      // the app name.
+      var fileName = Process.GetCurrentProcess().MainModule?.ModuleName ?? "<app>";
+      var usage = $"Usage: {fileName}";
+      Console.WriteLine(usage);
+      var usageLenWithPadding = usage.Length + 2;
+      foreach (var rule in _rules.Where(r => r is RequiredCommandlineArgumentRule))
+      {
+        WriteUsageRule(rule, usageLenWithPadding);
+      }
+      foreach (var rule in _rules.Where( r => r is OptionalCommandlineArgumentRule ))
+      {
+        WriteUsageRule(rule, usageLenWithPadding);
+      }
+      foreach (var rule in _rules.Where(r => r is HelpCommandlineArgumentRule))
+      {
+        WriteUsageRule(rule, usageLenWithPadding);
+      }
+    }
+
+    private void WriteUsageRule(ICommandlineArgumentRule rule, in int usageLenWithPadding)
+    {
+      // build the string
+      var line = FlattenKeys(rule.Keys, true);
+      if (rule.DefaultValue != null)
+      {
+        line = $"{line}={rule.DefaultValue}";
+      }
+      if (!rule.IsRequired)
+      {
+        line = $"[{line}]";
+      }
+
+      line = (new string(' ', usageLenWithPadding)) + line;
+      Console.WriteLine(line);
     }
 
     /// <summary>
@@ -53,10 +112,13 @@ namespace myoddweb.commandlineparser.OutputFormatter
     /// For example, []{ "h", "help"} would become "h, help"
     /// </summary>
     /// <param name="ruleKeys"></param>
+    /// <param name="addLeadingPattern"></param>
     /// <returns></returns>
-    private static string FlattenKeys(IEnumerable<string> ruleKeys)
+    private string FlattenKeys(ICollection<string> ruleKeys, bool addLeadingPattern )
     {
-      return string.Join(", ", ruleKeys);
+      var subKeys = new List<string>( ruleKeys.Count);
+      subKeys.AddRange(ruleKeys.Select(ruleKey => addLeadingPattern ? $"{_leadingPattern}{ruleKey}" : ruleKey));
+      return string.Join( ", ", subKeys);
     }
   }
 }