+++ src/myoddweb.commandlineparser/Rules/OptionalCommandlineArgumentRule.cs
@@ -24,13 +24,13 @@ namespace myoddweb.commandlineparser.Rules
 {
   public class OptionalCommandlineArgumentRule : CommandlineArgumentRule
   {
-    public OptionalCommandlineArgumentRule(string key, string description = null) :
-      this( new []{key}, description )
+    public OptionalCommandlineArgumentRule(string key) :
+      this(new[] { key })
     {
     }
 
     public OptionalCommandlineArgumentRule(string key, string defaultValue, string description = null) :
-      this( new []{key}, description, defaultValue)
+      this( new []{key}, defaultValue, description )
     {
       if (null == defaultValue)
       {
@@ -38,8 +38,8 @@ namespace myoddweb.commandlineparser.Rules
       }
     }
 
-    public OptionalCommandlineArgumentRule(IEnumerable<string> keys, string description = null) :
-      base( keys, false, description, null)
+    public OptionalCommandlineArgumentRule(IEnumerable<string> keys ) :
+      base(keys, false, null, null)
     {
     }